pyproject.toml
[project]
name = "xiaoai-desktop"
version = "0.1.0"
description = "基于 MQTT 的小爱同学桌面控制中心"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PySide6>=6.5",
  "paho-mqtt>=1.6,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
build = [
  "pyinstaller>=6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
